pyproject.toml
[tool.poetry]
name = "bookwriter-openai"
version = "0.1.0"
description = "You give the subject, I return the book"
authors = ["Thiago MadPin"]
license = "MIT"
readme = "README.md"
packages = [{include = "bookwriter_openai"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.6"
nltk = "^3.8.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.3.0"}
flake8 = "^6.0.0"
isort = "^5.12.0"
pip-tools = "^6.13.0"
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
